Refactored Gome interface
